move parse k8s log filename to meta pkg
