+++ decoder/cri.go
@@ -3,7 +3,6 @@ package decoder
 import (
 	"bytes"
 	"fmt"
-	"strings"
 )
 
 const (
@@ -68,37 +67,4 @@ func DecodeCRI(data []byte) (row CRIRow, _ error) {
 	row.Log = log
 
 	return row, nil
-}
-
-type CRIMetaInformation struct {
-	namespace     string
-	podName       string
-	containerName string
-	containerID   string
-}
-
-func NewCRIMetaInformation(fullFilename string) CRIMetaInformation {
-	lastSlash := strings.LastIndexByte(fullFilename, '/')
-	filename := fullFilename[lastSlash+1 : len(fullFilename)-4]
-	underscore := strings.IndexByte(filename, '_')
-	pod := filename[:underscore]
-	filename = filename[underscore+1:]
-	underscore = strings.IndexByte(filename, '_')
-	ns := filename[:underscore]
-	filename = filename[underscore+1:]
-	container := filename[:len(filename)-65]
-	cid := filename[len(filename)-64:]
-
-	return CRIMetaInformation{
-		ns, pod, container, cid,
-	}
-}
-
-func (m CRIMetaInformation) GetData() map[string]any {
-	return map[string]any{
-		"pod":          m.podName,
-		"namespace":    m.namespace,
-		"container":    m.containerName,
-		"container_id": m.containerID,
-	}
 }
