+++ decoder/cri_test.go
@@ -57,21 +57,3 @@ func TestCRIJoined3Lines(t *testing.T) {
 	assert.Equal(t, `{"level":"warn","ts":"2024-05-22T06:39:29.230Z"}\n`, string(row.Log))
 	assert.Equal(t, false, row.IsPartial)
 }
-
-func TestParseLogFilename(t *testing.T) {
-	metaInfo := NewCRIMetaInformation("/k8s-logs/advanced-logs-checker-1566485760-trtrq_sre_duty-bot-4e0301b633eaa2bfdcafdeba59ba0c72a3815911a6a820bf273534b0f32d98e0.log")
-
-	assert.Equal(t, "sre", metaInfo.namespace)
-	assert.Equal(t, "advanced-logs-checker-1566485760-trtrq", metaInfo.podName)
-	assert.Equal(t, "duty-bot", metaInfo.containerName)
-	assert.Equal(t, "4e0301b633eaa2bfdcafdeba59ba0c72a3815911a6a820bf273534b0f32d98e0", metaInfo.containerID)
-}
-
-func TestParseLogFilenameError(t *testing.T) {
-	metaInfo := NewCRIMetaInformation("web-logs-filed")
-
-	assert.Equal(t, "sre", metaInfo.namespace)
-	assert.Equal(t, "advanced-logs-checker-1566485760-trtrq", metaInfo.podName)
-	assert.Equal(t, "duty-bot", metaInfo.containerName)
-	assert.Equal(t, "4e0301b633eaa2bfdcafdeba59ba0c72a3815911a6a820bf273534b0f32d98e0", metaInfo.containerID)
-}
